requirements.txt
praw==7.6.0
PyMySQL==1.0.2
pmaw~=2.1.3
requests~=2.27.1
prawcore~=2.3.0
Pillow~=9.2.0
ImageHash~=4.2.1
opencv-python~=4.6.0.66
numpy~=1.23.1